Break discovery scanner import cycle
